refactor: websocket controller now is a production tool, and it was split from it's own class
