+++ unity-renderer/Assets/Scenes/InitialScene.unity
@@ -274,12 +274,6 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: a9c131371473347fab2b6a6dbfe8c13b, type: 3}
---- !u!1 &287244851 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 6401575379531892522, guid: a9c131371473347fab2b6a6dbfe8c13b,
-    type: 3}
-  m_PrefabInstance: {fileID: 287244850}
-  m_PrefabAsset: {fileID: 0}
 --- !u!1 &312495477
 GameObject:
   m_ObjectHideFlags: 0
@@ -401,18 +395,6 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: c96c5543d89960749bdf3025e5fc122b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!114 &322303339 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 1629508456149027013, guid: b387035aa46399445bdc7951a41d5213,
-    type: 3}
-  m_PrefabInstance: {fileID: 322303337}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4c1e443212437a148b1a6f54aea2ab71, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1 &491237585
@@ -1081,7 +1063,7 @@ GameObject:
   - component: {fileID: 1444882337}
   - component: {fileID: 1444882336}
   m_Layer: 0
-  m_Name: WSSController
+  m_Name: DebugParameters
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1096,14 +1078,9 @@ MonoBehaviour:
   m_GameObject: {fileID: 1444882335}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8b5e6e1eec417844b92edd6ca2f8a2b2, type: 3}
+  m_Script: {fileID: 11500000, guid: 73b83adac0a4ddd039a0a1c9c2a9a163, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  renderingController: {fileID: 322303339}
-  characterController: {fileID: 1989862556}
-  cameraController: {fileID: 1608366483}
-  bridgesGameObject: {fileID: 287244851}
-  wssServerUrl: ws://localhost:4999/
   openBrowserWhenStart: 1
   useCustomContentServer: 0
   customContentServerUrl: http://localhost:1338/
@@ -1209,18 +1186,6 @@ Transform:
     type: 3}
   m_PrefabInstance: {fileID: 326638024064006570}
   m_PrefabAsset: {fileID: 0}
---- !u!114 &1989862556 stripped
-MonoBehaviour:
-  m_CorrespondingSourceObject: {fileID: 7249126750621343183, guid: b0965b55b892d438da492708ea09799c,
-    type: 3}
-  m_PrefabInstance: {fileID: 326638024064006570}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: d1d0821bf5c3e3f47ba67e483803236d, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!114 &2021276774 stripped
 MonoBehaviour:
   m_CorrespondingSourceObject: {fileID: 1312815038965150261, guid: b0965b55b892d438da492708ea09799c,
