+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/Scene/Main/Main.asmdef
@@ -6,7 +6,6 @@
         "GUID:86e5747efa9128344bd8e4754db51d3b",
         "GUID:28f74c468a54948bfa9f625c5d428f56",
         "GUID:e945202cffc30cc479229aef4577a74c",
-        "GUID:5b43b30b2a81b9df795c669fe8f3d84c",
         "GUID:b1087c5731ff68448a0a9c625bb7e52d",
         "GUID:cc6bfabbfe87b7949aa248893f89ce37",
         "GUID:3f13bd1a926f03a4cab988f06a80c1fd",
@@ -19,7 +18,10 @@
         "GUID:2eaa9b64aed3308028df3c2f56a3a471",
         "GUID:c27391cd3f3b5aeb4b4c45372ede8043",
         "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
-        "GUID:873fc6254de64e84ba87ff26c2bbafca"
+        "GUID:873fc6254de64e84ba87ff26c2bbafca",
+        "GUID:a166f06298ffb5fdaaf744ff18e2a5b8",
+        "GUID:9ee23d03bb7793ad3a3c66f4fa65e722",
+        "GUID:c6e3ac0a2748a14e2b281e440c082eb6"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
