+++ unity-renderer/Assets/Scripts/MainScripts/DCL/DataStore/DataStore.cs
@@ -29,6 +29,7 @@ namespace DCL
         public readonly BaseVariable<NFTPromptModel> onOpenNFTPrompt = new BaseVariable<NFTPromptModel>();
         public readonly DataStore_AvatarsLOD avatarsLOD = new DataStore_AvatarsLOD();
         public readonly DataStore_VirtualAudioMixer virtualAudioMixer = new DataStore_VirtualAudioMixer();
+        public readonly DataStore_WSCommunication wsCommunication = new DataStore_WSCommunication();
 
         public class DataStore_BuilderInWorld
         {
@@ -91,6 +92,18 @@ namespace DCL
             public readonly BaseVariable<float> voiceChatVolume = new BaseVariable<float>(1f);
             public readonly BaseVariable<float> uiSFXVolume = new BaseVariable<float>(1f);
             public readonly BaseVariable<float> avatarSFXVolume = new BaseVariable<float>(1f);
+        }
+
+        public class DataStore_WSCommunication
+        {
+            [System.NonSerialized]
+            public string wssServerUrl = "ws://localhost:5000/";
+            
+            [System.NonSerialized]
+            public string wssServiceId = "dcl";
+            
+            public readonly BaseVariable<bool> communicationEstablished = new BaseVariable<bool>();
+            public readonly BaseVariable<bool> communicationReady = new BaseVariable<bool>();
         }
     }
 }
