--- unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/EntryPoint_World/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("EntryPoint_World_Tests")]
