--- unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/EntryPoint_World/Tests/Tests.asmdef
@@ -1,27 +1,0 @@
-{
-    "name": "EntryPoint_World_Tests",
-    "rootNamespace": "",
-    "references": [
-        "UnityEngine.TestRunner",
-        "UnityEditor.TestRunner",
-        "MainScripts",
-        "Protocol",
-        "QueuedSceneMessage",
-        "MessagingInterfaces",
-        "EntryPoint_World"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": true,
-    "precompiledReferences": [
-        "nunit.framework.dll",
-        "Newtonsoft.Json.dll"
-    ],
-    "autoReferenced": false,
-    "defineConstraints": [
-        "UNITY_INCLUDE_TESTS"
-    ],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
