+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: 5b43b30b2a81b9df795c669fe8f3d84c
-AssemblyDefinitionImporter:
+guid: 4b0ecf55542130a7a93607e41532035e
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
