+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/IKernelCommunication.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.PlayerLoop;
+
+public class IKernelCommunication
+{
+    public IKernelCommunication()
+    {
+        
+    }
+
+    public virtual void Update()
+    {
+        
+    }
+}
