+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/IKernelCommunication.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: cd2bccef28915644fb7b3100d04e85a2
+guid: f4292d337df1e4baf95f1231625f4282
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
