+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/KernelCommunication.asmdef
@@ -1,0 +1,3 @@
+{
+	"name": "KernelCommunication"
+}
