+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/KernelCommunication.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: dc74831c7fd753a4fa0aa577ab09d90d
-folderAsset: yes
-DefaultImporter:
+guid: a166f06298ffb5fdaaf744ff18e2a5b8
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
