+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3068b77539728853b88ff28f436616c9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
