+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/AssemblyInfo.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8b5e6e1eec417844b92edd6ca2f8a2b2
+guid: b253ef72d84a931d8b8b3af790be1ae2
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
