+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/NativeBridgeCommunication.asmdef
@@ -1,11 +1,12 @@
 {
-    "name": "EntryPoint_World",
+    "name": "NativeBridgeCommunication",
     "rootNamespace": "",
     "references": [
         "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
         "GUID:a881b57670b1d2747a6d7f9e32b63230",
         "GUID:132013f160565434195ba5960832b346",
-        "GUID:8c63f342fdbca7e1fbf6a616655e1c54"
+        "GUID:8c63f342fdbca7e1fbf6a616655e1c54",
+        "GUID:a166f06298ffb5fdaaf744ff18e2a5b8"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
