+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/NativeBridgeCommunication.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 9ee23d03bb7793ad3a3c66f4fa65e722
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
