+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/NativeBridgeCommunication.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e8c9bce799cb62469a73551ef026f93
+guid: 309d2aa5e2ee4f65a97c7c69a78fe393
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
