+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/Tests.meta
@@ -1,6 +1,7 @@
 fileFormatVersion: 2
-guid: e03ac062342aa8347be274ed5965e374
-AssemblyDefinitionImporter:
+guid: b94d028feda1124ea98152684f25df7a
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
