+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/Tests/NativeBridgeCommunicationShould.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a143f7d5b24952649bf93d8b78f67212
+guid: c9ec7895898dc95b9b7e3db150c74f65
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
