+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/NativeBridgeCommunication/Tests/Tests.asmdef
@@ -1,0 +1,28 @@
+{
+    "name": "NativeBridgeCommunicationTests",
+    "rootNamespace": "",
+    "references": [
+        "GUID:27619889b8ba8c24980f49ee34dbb44a",
+        "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:e555144732b726c4cba5b0f6337fea75",
+        "GUID:a881b57670b1d2747a6d7f9e32b63230",
+        "GUID:132013f160565434195ba5960832b346",
+        "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
+        "GUID:9ee23d03bb7793ad3a3c66f4fa65e722",
+        "GUID:a166f06298ffb5fdaaf744ff18e2a5b8"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": true,
+    "precompiledReferences": [
+        "nunit.framework.dll",
+        "Newtonsoft.Json.dll"
+    ],
+    "autoReferenced": false,
+    "defineConstraints": [
+        "UNITY_INCLUDE_TESTS"
+    ],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
