+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 87d6a9a295e29cf49bdbba972e0e8dc9
+guid: 520a1a71df9a1636abcef04e98668c37
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
