+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication/Tests.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4199cca988b4b62389a3dd80106a611d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
