+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication/Tests/Tests.asmdef
@@ -8,14 +8,14 @@
         "GUID:ef84d7ce90ba34f07be26ace428e9bc8",
         "GUID:27619889b8ba8c24980f49ee34dbb44a",
         "GUID:0acc523941302664db1f4e527237feb3",
-        "GUID:d7ced3c8f2800ed40a132e6df054755e",
         "GUID:c9fbb75c15a4cd042b731877ede4a5c1",
         "GUID:97d8897529779cb49bebd400c7f402a6",
         "GUID:fbcc413e192ef9048811d47ab0aca0c0",
         "GUID:2c6e13fc1d2a309239fafe1b0de41193",
         "GUID:9aa18dd7d97b0419e933b45b56d4ccca",
         "GUID:2eaa9b64aed3308028df3c2f56a3a471",
-        "GUID:c27391cd3f3b5aeb4b4c45372ede8043"
+        "GUID:c27391cd3f3b5aeb4b4c45372ede8043",
+        "GUID:c6e3ac0a2748a14e2b281e440c082eb6"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
