+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication/WebSocketCommunication.asmdef
@@ -1,0 +1,22 @@
+{
+    "name": "WebSocketCommunication",
+    "rootNamespace": "",
+    "references": [
+        "GUID:7deefc26eb4ef824586f2f51caa3b0e7",
+        "GUID:4973650d2444c4561a15d50f24d91cd9",
+        "GUID:a166f06298ffb5fdaaf744ff18e2a5b8",
+        "GUID:28f74c468a54948bfa9f625c5d428f56",
+        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
+        "GUID:97d8897529779cb49bebd400c7f402a6",
+        "GUID:3047fd2f05bcaae498a99a3f9321f9a8"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
