+++ unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication/WebSocketCommunication.asmdef.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: cb0e2599aaea8a64cbb18a7eec726ec9
-folderAsset: yes
-DefaultImporter:
+guid: c6e3ac0a2748a14e2b281e440c082eb6
+AssemblyDefinitionImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
