+++ unity-renderer/Assets/Scripts/MainScripts/Debugging/DebugParameters/DebugParameters.asmdef
@@ -1,0 +1,17 @@
+{
+    "name": "DebugParameters",
+    "rootNamespace": "",
+    "references": [
+        "GUID:e555144732b726c4cba5b0f6337fea75",
+        "GUID:28f74c468a54948bfa9f625c5d428f56"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
