+++ unity-renderer/Assets/Scripts/MainScripts/Debugging/DebugParameters/DebugParameters.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 73b83adac0a4ddd039a0a1c9c2a9a163
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
