--- unity-renderer/Assets/Scripts/MainScripts/Debugging/WSS/WSSController.asmdef
@@ -1,45 +1,0 @@
-{
-    "name": "WSSController",
-    "rootNamespace": "",
-    "references": [
-        "GUID:e555144732b726c4cba5b0f6337fea75",
-        "GUID:7deefc26eb4ef824586f2f51caa3b0e7",
-        "GUID:1de3566cccb280f4a982c59ad0d08c96",
-        "GUID:4973650d2444c4561a15d50f24d91cd9",
-        "GUID:7ac9f9c835ec1084ab35e3f9b176cf1e",
-        "GUID:0e967802b778d404eac1ca5ea340e290",
-        "GUID:31d8dc012a18a4a6da741a0fd9c390e0",
-        "GUID:1a483cdd86db6e641ad77c7046dd035f",
-        "GUID:8e14059e964ca49aeb55d0e7b2b14c3c",
-        "GUID:7240b5b323e5f4bde8392d35668ad0ae",
-        "GUID:eb95615676ca6984687c5ae7774d6309",
-        "GUID:ba9b721ce33233c4da7ff4ef30d25b59",
-        "GUID:70aa308435753374584c7061e4b89ef2",
-        "GUID:b8662e798190944f4b638fbeb48e84f0",
-        "GUID:086c21f5b0c2a6a488e34b5e5da4eb40",
-        "GUID:4720e174f2805c74bb7aa94cc8bb5bf8",
-        "GUID:28f74c468a54948bfa9f625c5d428f56",
-        "GUID:0f686f1bcb4670a4bb2929a41daaf716",
-        "GUID:6dacc6cde5056a34c9488c309ca0a18a",
-        "GUID:a1e1473c53889c2489c0822a9dd5fa33",
-        "GUID:9b06d63e1a13bc34c995683595d8d849",
-        "GUID:81a6b1347b86bfe4aa02ed089ff533c1",
-        "GUID:97d8897529779cb49bebd400c7f402a6",
-        "GUID:335b5dc19a4772e43a9cbc73799fc914",
-        "GUID:fbcc413e192ef9048811d47ab0aca0c0",
-        "GUID:3047fd2f05bcaae498a99a3f9321f9a8",
-        "GUID:d7f1765642f0c22cda30addd0275b5e1",
-        "GUID:3e062a4ccaf3f77919ca87fd7ad3743f",
-        "GUID:3bdb7fe910cce8d4888137cded3ba4a2",
-        "GUID:e0a7432b43a6cbc4a8db4e39c83ee95e"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
